eval/harness/requirements.txt
PyYAML==6.0.2
requests==2.32.3
SQLAlchemy==2.0.32
psycopg[binary]==3.2.1

